keypoint_detector/requirements.txt
opencv-python==4.8.1.78
pyyaml==6.0.1
minio==7.2.7
psycopg2-binary==2.9.9
git+https://github.com/HumanSignal/label-studio-sdk@master
jupyter==1.0.0
matplotlib
tensorflow==2.15.0